requirements.txt
flappy-bird-gymnasium
tensorflow
PyYAML
--index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio